requirements.txt
# Principal Narrative API Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# YAML parsing for frontmatter
pyyaml>=6.0.1

# HTTP client
httpx>=0.25.0

# Development
python-multipart>=0.0.6

# LLM Integration
anthropic>=0.39.0

# Vector store for semantic search
chromadb>=0.4.0

# Slack integration (optional)
slack-sdk>=3.23.0

# GitHub integration (optional)
PyGithub>=2.1.0

# Authentication
PyJWT>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Redis for caching (optional)
# redis>=5.0.0
